Rename python_reconfigure_installed_package.py to python_reconfigure_an_installed_package.py
This script takes the name of the package as a command-line argument and uses subprocess.run() to execute the dpkg-reconfigure command. It then checks the exit status of the command and prints a success message if the reconfiguration is successful or an error message if it fails.

To use this script:

Save the script to a Python file, for example, reconfigure_package.py.
Run the script with the package name as an argument, like so: python3 reconfigure_package.py <package_name>.
Replace <package_name> with the name of the package you want to reconfigure. For example, to reconfigure the mysql-server package, you would run: python3 reconfigure_package.py mysql-server.