Create python_get_information_about_specific_package.py
This script takes the name of the package as a command-line argument and uses subprocess.run() to execute the apt-cache show command to get information about the package. It then prints the information to the console.

To use this script:

Save the script to a Python file, for example, get_package_information.py.
Run the script with the package name as an argument, like so: python3 get_package_information.py <package_name>.
Replace <package_name> with the name of the package you want to get information about. For example, to get information about the apache2 package, you would run: python3 get_package_information.py apache2.