Create python_Removing_Packages.py
This script takes the name of the package as a command-line argument and uses subprocess.run() to execute the apt remove command to remove the package. It then prints a success message if the package is removed successfully or an error message if it fails.

To use this script:

Save the script to a Python file, for example, remove_package.py.
Run the script with the package name as an argument, like so: python3 remove_package.py <package_name>.
Replace <package_name> with the name of the package you want to remove. For example, to remove the apache2 package, you would run: python3 remove_package.py apache2.