Create python_shutdown_and_restart_the_system.py
This script takes the action (shutdown or restart) as a command-line argument and uses subprocess.run() to execute the shutdown command accordingly. It then prints a success message if the system is shutdown or restarted successfully or an error message if it fails.

To use this script:

Save the script to a Python file, for example, shutdown_system.py.
Run the script with the action (shutdown or restart) as an argument, like so: python3 shutdown_system.py <action>.
Replace <action> with either shutdown or restart. For example, to shutdown the system, you would run: python3 shutdown_system.py shutdown.