Create python_create_archive.py
This script takes the archive name and a list of files or directories to include in the archive as command-line arguments. It then constructs and executes the tar command to create the archive using the specified files or directories.

To use this script:

Save the script to a Python file, for example, create_archive.py.
Run the script using Python with the archive name and list of files/directories to include as arguments:
php
Copy code
python3 create_archive.py <archive_name> <file1> <file2> ...
Replace <archive_name> with the desired name of the archive, and <file1>, <file2>, etc., with the paths of the files or directories you want to include in the archive.