Create python_extract_archive.py
This script takes the archive name and an optional destination directory as command-line arguments. It then constructs and executes the tar command to extract the archive to the specified destination directory.

To use this script:

Save the script to a Python file, for example, extract_archive.py.
Run the script using Python with the archive name and optional destination directory as arguments:
css
Copy code
python3 extract_archive.py <archive_name> [destination]
Replace <archive_name> with the name of the archive you want to extract, and [destination] with the optional destination directory where you want to extract the archive. If no destination directory is provided, the files will be extracted to the current directory.